Extract base class for lottie player views
Creating the player for the file rendered in a canvas and destroying
it again is the same for the thumbnail, the file meta data preview and
the box of the positioning dialog. What differs is playback and how
each of them reacts to the animation having been read.
