+++ entry_types/scrolled/package/src/contentElements/lottieAnimation/editor/views/LottieFileThumbnailView.js
@@ -1,36 +1,19 @@
-import Marionette from 'backbone.marionette';
-
-import {cssModulesUtils} from 'pageflow/ui';
-
-import {DotLottie} from '../../dotLottie';
+import {LottieFilePlayerView} from './LottieFilePlayerView';
 
 import styles from './LottieFileThumbnailView.module.css';
 
-export const LottieFileThumbnailView = Marionette.ItemView.extend({
+export const LottieFileThumbnailView = LottieFilePlayerView.extend({
   template: () => `<canvas class="${styles.canvas}"></canvas>`,
   className: styles.thumbnail,
 
-  ui: cssModulesUtils.ui(styles, 'canvas'),
-
-  onRender: function() {
-    this.player = new DotLottie({
-      canvas: this.ui.canvas[0],
-      src: this.model.get('original_url'),
-      autoplay: false,
-      renderConfig: {autoResize: true}
-    });
-
-    this.player.addEventListener('load', this.seekToLastFrame.bind(this));
+  playerOptions: function() {
+    return {autoplay: false};
   },
 
   // Animations commonly build up their scene over time, which would
   // leave the thumbnail close to blank on the first frame the player
   // draws once it has loaded the file.
-  seekToLastFrame: function() {
+  onAnimationLoad: function() {
     this.player.setFrame(this.player.totalFrames - 1);
-  },
-
-  onClose: function() {
-    this.player.destroy();
   }
 });
